Lift the two largest per-kind method groups out of SceneDocument
SceneDocument had 144 public members, 79 of them specific to a single node
kind, in a 3,037-line module. That is most of what makes adding a node
kind a 44-file change, and it is the half of QA item C3 that is real -
the other half, "kind-specific fields left on SceneNode", turned out on
measurement to be shared fields that belong there (see #400, #402).

backend/domain/ already decomposes its CROSS-CUTTING concerns this way:
SceneDocument is composed from BranchOps, GroupOps, LayoutOps and
CommandOps. This extends the same idiom to the PER-KIND ones, starting
with the two largest groups:

    backend/domain/nodes_code_review.py   CodeReviewOps    9 methods
    backend/domain/nodes_gitlink.py       GitlinkOps      11 methods

    graph.py                3,037 -> 2,591 lines
    methods in graph.py        93 -> 73
    SceneDocument public API  144 -> 144  (unchanged, by design)

Both mixins need only what SceneDocumentParts already declares - nodes,
connect, _counter - so they inherit it and stay type-checkable in
isolation like their cross-cutting siblings, and
tests/test_mixin_declaration_bases.py covers them automatically through
the MRO rather than by being told about them.

This is a relocation, not a rewrite. Every one of the 20 moved methods was
verified AST-identical to its pre-move version (ast.unparse round-trip
compared against `git show HEAD:backend/domain/graph.py`), and every one
is still reachable on SceneDocument. store_code_review_diff's own
_bundle_int helper moved with the methods that use it.

Two kinds, not all fifteen: this proves the extraction on the largest
groups and leaves the pattern for the rest to follow one at a time. The
remaining thirteen are 13-128 lines each.

Test plan:
- 20/20 moved methods AST-identical; MRO is
  SceneDocument -> Branch/Group/Layout/Command/CodeReview/Gitlink Ops ->
  SceneDocumentParts -> object.
- tests/test_domain_purity.py green: neither new module imports anything
  the domain layer may not.
- Full suite: 3224 passed, 19 skipped. ruff and mypy clean.
